reset iat in generate_auth in order to make refreshing auth easier
because we're promoting calling `firebase.auth = nil` to refresh the
token. to make this work, we need to reset iat in generate_auth.
but really this is a bit bad because setting auth or calling
generate_auth would affect iat, this is definitely not expected!

but how do we keep it backward compatible while making this more sense?
any good idea?